Update headless.md for country definition
raspi-config gives an error about being unable to communicate with wpa-supplicant due to the country not being defined.  The current documentation on both headless and the wifi setup never mention the country option.  The Wifi page indicates that you should edit the existing wpa-supplicant.conf, which may already contain this.